edge/requirements.txt
numpy<2.0.0
setproctitle
opencv-python
easyocr
notebook
ipykernel
degirum
degirum_tools
degirum_cli
picamera2
pillow<11
matplotlib
Flask
Flask-SocketIO
websockets
python-dotenv
pandas
matplotlib
requests
sqlite3
libcamera
difflib
eventlet